@@ -5,21 +5,19 @@ namespace Dena\IranPayment\Traits;
 trait IranPaymentDatabase
 {
     /**
-     * Table Name variable
+     * IranPayment Table Name variable
      *
      * @var string
      */
-    private $iranpayment_table = 'iranpayment_transactions';
+    private string $iranpayment_table = 'iranpayment_transactions';
 
     /**
-     * Get Table Name function
+     * Get IranPayment Table Name function
      *
      * @return string
      */
     public function getTable(): string
     {
-        $this->table = config('iranpayment.table', $this->iranpayment_table);
-
-        return $this->table;
+        return $this->table = app('config')->get('iranpayment.table', $this->iranpayment_table);
     }
 }
