@@ -4,6 +4,7 @@ namespace Dena\IranPayment\Exceptions;
 
 use Dena\IranPayment\Exceptions\IranPaymentException;
 
+use Exception;
 use Throwable;
 
 class GatewayException extends IranPaymentException
@@ -13,18 +14,27 @@ class GatewayException extends IranPaymentException
         parent::__construct($message, $code, $previous);
     }
 
-    public static function unknownResponse()
+    public static function unknownResponse(string $response = null): self
     {
-        return new self('پاسخ ناشناخته!');
+        return new self(
+            'پاسخ ناشناخته!',
+            500,
+            new Exception($response)
+        );
     }
 
-    public static function inconsistentResponse()
+    public static function inconsistentResponse(): self
     {
         return new self('اطلاعات دریافتی با پایگاه‌داده همخوانی ندارند!');
     }
 
-    public static function connectionProblem()
+    public static function connectionProblem(Throwable $previous = null): self
     {
-        return new self('اشکالی در اتصال به درگاه پرداخت پیش آمده است!', 503);
+        return new self('در اتصال به درگاه پرداخت اشکالی پیش آمده است!', 503, $previous);
+    }
+
+    public static function notSupportedMethod(): self
+    {
+        return new self('درگاه پرداخت از این متد پشتیبانی نمی‌کند!');
     }
 }
